requirements.txt
fastapi==0.123.0
uvicorn[standard]==0.30.0
pydantic==2.12.5
python-multipart
Pillow
numpy